mechanical move of all localfs dependencies on os.FileInfo to a separate file
